💥refactor: NetworkAPI 수정
